+++ lib/iolog/iolog_filter.c
@@ -29,19 +29,19 @@
 #ifdef HAVE_STDBOOL_H
 # include <stdbool.h>
 #else
-# include "compat/stdbool.h"
+# include <compat/stdbool.h>
 #endif
 #include <regex.h>
 #include <string.h>
 #include <time.h>
 
-#include "sudo_compat.h"
-#include "sudo_debug.h"
-#include "sudo_fatal.h"
-#include "sudo_gettext.h"
-#include "sudo_iolog.h"
-#include "sudo_queue.h"
-#include "sudo_util.h"
+#include <sudo_compat.h>
+#include <sudo_debug.h>
+#include <sudo_fatal.h>
+#include <sudo_gettext.h>
+#include <sudo_iolog.h>
+#include <sudo_queue.h>
+#include <sudo_util.h>
 
 struct pwfilt_regex {
     TAILQ_ENTRY(pwfilt_regex) entries;
@@ -177,7 +177,7 @@ iolog_pwfilt_remove(void *vhandle, const char *pattern)
  */
 bool
 iolog_pwfilt_run(void *vhandle, int event, const char *buf,
-    unsigned int len, char **newbuf)
+    size_t len, char **newbuf)
 {
     struct pwfilt_handle *handle = vhandle;
     struct pwfilt_regex *filt;
