+++ logsrvd/logsrv_util.c
@@ -29,7 +29,7 @@
 #ifdef HAVE_STDBOOL_H
 # include <stdbool.h>
 #else
-# include "compat/stdbool.h"
+# include <compat/stdbool.h>
 #endif /* HAVE_STDBOOL_H */
 #if defined(HAVE_STDINT_H)
 # include <stdint.h>
@@ -42,12 +42,12 @@
 #include <time.h>
 #include <unistd.h>
 
-#include "sudo_compat.h"
-#include "sudo_debug.h"
-#include "sudo_fatal.h"
-#include "sudo_gettext.h"
-#include "sudo_iolog.h"
-#include "sudo_util.h"
+#include <sudo_compat.h>
+#include <sudo_debug.h>
+#include <sudo_fatal.h>
+#include <sudo_gettext.h>
+#include <sudo_iolog.h>
+#include <sudo_util.h>
 
 #include "logsrv_util.h"
 
@@ -55,16 +55,16 @@
  * Expand buf as needed or just reset it.
  */
 bool
-expand_buf(struct connection_buffer *buf, unsigned int needed)
+expand_buf(struct connection_buffer *buf, size_t needed)
 {
     void *newdata;
     debug_decl(expand_buf, SUDO_DEBUG_UTIL);
 
     if (buf->size < needed) {
 	/* Expand buffer. */
-	const unsigned int newsize = sudo_pow2_roundup(needed);
+	const size_t newsize = sudo_pow2_roundup(needed);
 	sudo_debug_printf(SUDO_DEBUG_INFO|SUDO_DEBUG_LINENO,
-	    "expanding buffer from %u to %u", buf->size, newsize);
+	    "expanding buffer from %zu to %zu", buf->size, newsize);
 	if (newsize < needed) {
 	    /* overflow */
 	    errno = ENOMEM;
@@ -166,7 +166,7 @@ iolog_seekto(int iolog_dir_fd, const char *iolog_path,
 		    iolog_fd_to_name(timing.event));
 		goto bad;
 	    }
-	    pos = iolog_seek(&iolog_files[timing.event], timing.u.nbytes,
+	    pos = iolog_seek(&iolog_files[timing.event], (off_t)timing.u.nbytes,
 		SEEK_CUR);
 	    if (pos == -1) {
 		sudo_warn(U_("%s/%s: unable to seek forward %zu"), iolog_path,
