+++ plugins/python/pyhelpers.h
@@ -26,8 +26,8 @@
 #undef SIZEOF_TIME_T
 
 #include <config.h>
-#include "sudo_compat.h"
-#include "sudo_plugin.h"
+#include <sudo_compat.h>
+#include <sudo_plugin.h>
 
 #include "pyhelpers_cpychecker.h"
 
@@ -85,18 +85,19 @@ CPYCHECKER_RETURNS_BORROWED_REF
 PyObject *py_tuple_get(PyObject *py_tuple, Py_ssize_t index, PyTypeObject *expected_type);
 
 PyObject *py_object_get_optional_attr(PyObject *py_object, const char *attr, PyObject *py_default);
-long long py_object_get_optional_attr_number(PyObject *py_object, const char *attr_name);
+long py_object_get_optional_attr_number(PyObject *py_object, const char *attr_name);
 const char *py_object_get_optional_attr_string(PyObject *py_object, const char *attr_name);
 
-void py_object_set_attr_number(PyObject *py_object, const char *attr_name, long long number);
+void py_object_set_attr_number(PyObject *py_object, const char *attr_name, long number);
 void py_object_set_attr_string(PyObject *py_object, const char *attr_name, const char *value);
 
 PyObject *py_create_version(unsigned int version);
 
 void py_debug_python_call(const char *class_name, const char *function_name,
-                          PyObject *py_args, PyObject *py_kwargs, int subsystem_id);
+                          PyObject *py_args, PyObject *py_kwargs,
+                          unsigned int subsystem_id);
 void py_debug_python_result(const char *class_name, const char *function_name,
-                            PyObject *py_args, int subsystem_id);
+                            PyObject *py_args, unsigned int subsystem_id);
 
 void str_array_free(char ***array);
