+++ src/env_hooks.c
@@ -27,9 +27,9 @@
 #include <string.h>
 #include <errno.h>
 
-#include "sudo.h"
-#include "sudo_plugin.h"
-#include "sudo_dso.h"
+#include <sudo.h>
+#include <sudo_plugin.h>
+#include <sudo_dso.h>
 
 extern char **environ;		/* global environment pointer */
 static char **priv_environ;	/* private environment pointer */
@@ -96,7 +96,7 @@ rpl_putenv(PUTENV_CONST char *string)
     }
 
     /* Look for existing entry. */
-    len = (equal - string) + 1;
+    len = (size_t)(equal - string) + 1;
     for (ep = environ; *ep != NULL; ep++) {
 	if (strncmp(string, *ep, len) == 0) {
 	    *ep = (char *)string;
